# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Aspectus API
  description: |
    Pandaria 生态统一身份与多租户管理服务。

    ## 认证

    **管理 API** 使用 Service Token：
    ```
    Authorization: Bearer {service_token}
    ```

    **自省端点** 使用 Service Token 认证调用方，subject token 通过 body 传递。

    **Auth 端点** (`/login`、`/register`、`/logout` 等) 无需认证。

    **OAuth2 端点** (`/authorize`, `/oauth/token`) 无需认证。

    ## Token 类型

    | 前缀 | 类型 |
    |------|------|
    | `pk_live_*` | API Key |
    | `eyJ*` | JWT |
    | `ot_*` | Opaque Token |
    | `rt_*` | Refresh Token |

    ## 错误格式

    所有错误使用 [RFC 7807](https://tools.ietf.org/html/rfc7807) Problem Details：
    ```json
    {
      "type": "https://aspectus.dev/errors/unauthorized",
      "title": "Unauthorized",
      "status": 401,
      "detail": "...",
      "instance": "/path"
    }
    ```

    `/introspect` 对无效 token 返回 `{"active": false}` (HTTP 200)，不返回错误。

  version: "0.9.0"
  contact:
    name: Aspectus Team
servers:
  - url: http://localhost:3100
    description: Development server
...
components:
  securitySchemes:
    ServiceToken:
      type: http
      scheme: bearer
      description: Service Token（每个 Project 一个）
...
  schemas:
...
    IntrospectResponse:
      type: object
      properties:
        active:
          type: boolean
        tenant_id:
          type: string
        user_id:
          type: string
        identity_type:
          type: string
          enum: [user, service_account]
        client_id:
          type: string
        scope:
          type: string
          description: 空格分隔的 scope 列表
        token_type:
          type: string
        token_format:
          type: string
          enum: [api_key, jwt, opaque]
        exp:
          type: integer
          format: int64
        quotas:
          type: object
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
...
paths:
...
  /introspect:
    post:
      tags: [Introspection]
      summary: Token 自省 (RFC 7662)
      security:
        - ServiceToken: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: 待验证的 token
                token_type_hint:
                  type: string
                  description: 可选。`api_key` | `jwt` | `opaque`
      responses:
        "200":
          description: 自省结果
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/IntrospectResponse"
                  - type: object
                    properties:
                      active:
                        type: boolean
                        enum: [false]
        "401":
          description: Service Token 无效
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /token:
    post:
      tags: [Introspection]
      summary: 签发 access token
      description: |
        使用 client_credentials 为 service account 或 machine client 签发
        JWT 或 Opaque token。
      security:
        - ServiceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenIssueRequest"
      responses:
        "200":
          description: Token 签发成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: client_secret 无效或 client_id/project 不匹配
  /token/revoke:
    post:
      tags: [Introspection]
      summary: 吊销 token
      description: |
        吊销 API Key 或 Opaque token。吊销后 `/introspect` 返回 `active: false`。
      security:
        - ServiceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenRevokeRequest"
      responses:
        "204":
          description: 吊销成功（幂等）

  # ── Tenants ──
  /tenants:
    post:
      tags: [Tenants]
      summary: 创建租户
      security:
        - ServiceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 128
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
    get:
      tags: [Tenants]
      summary: 列出所有租户
      security:
        - ServiceToken: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"
  /tenants/{id}:
    get:
      tags: [Tenants]
      summary: 查询租户
      security:
        - ServiceToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "404":
          $ref: "#/components/responses/NotFound"
  /tenants/{id}/quotas:
    put:
      tags: [Tenants]
      summary: 更新租户配额
      security:
        - ServiceToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: |
                以 project 为键的配额对象。每个 project 的值为 object，
                其中已知数值型字段必须是非负整数：monthly_tokens,
                max_concurrent_sessions, max_users, max_api_keys,
                requests_per_minute, monthly_requests。
              example:
                pandaria:
                  max_concurrent_sessions: 50
                tokencamp:
                  monthly_tokens: 10000000
      responses:
        "204":
          description: 更新成功
        "422":
          description: 配额 schema 不合法（未知 project 或数值型字段非整数）
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"

  # ── Service Accounts ──
  /service-accounts:
    post:
      tags: [Service Accounts]
      summary: 创建服务账号
      security:
        - ServiceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenant_id, label]
              properties:
                tenant_id:
                  type: string
                label:
                  type: string
                description:
                  type: string
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceAccount"
    get:
      tags: [Service Accounts]
      summary: 列出服务账号
      security:
        - ServiceToken: []
      parameters:
        - name: tenant_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceAccount"
  /service-accounts/{id}:
    get:
      tags: [Service Accounts]
      summary: 查询服务账号
      security:
        - ServiceToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceAccount"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Users ──
  /users:
    post:
      tags: [Users]
      summary: 创建用户
      security:
        - ServiceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenant_id, email, password]
              properties:
                tenant_id:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                display_name:
                  type: string
      responses:
        "201":
          description: 创建成功（不返回 password_hash）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    get:
      tags: [Users]
      summary: 列出用户
      security:
        - ServiceToken: []
      parameters:
        - name: tenant_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      tags: [Users]
      summary: 查询用户
      security:
        - ServiceToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{id}/suspend:
    put:
      tags: [Users]
      summary: 挂起/恢复用户
      security:
        - ServiceToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [suspended]
              properties:
                suspended:
                  type: boolean
      responses:
        "204":
          description: 操作成功
  /users/{id}/unlock:
    post:
      tags: [Users]
      summary: 解锁用户账户
      description: |
        清除登录失败计数并解除账户锁定。需要 admin service token。
      security:
        - ServiceToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: 解锁成功
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{id}/change-password:
    post:
      tags: [Users]
      summary: 自助修改密码
      description: |
        用户修改自己的登录密码，必须提供当前密码。该端点公开可访问，
        受 IP 级 rate limit 保护。
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password:
                  type: string
                new_password:
                  type: string
                  minLength: 8
      responses:
        "204":
          description: 修改成功
        "401":
          description: 当前密码错误或用户不存在/被挂起
        "422":
          description: 新密码过短
  /users/{id}/scopes:
    get:
      tags: [Users]
      summary: 查询用户有效 scopes
      description: 返回用户所有角色展开后合并去重的 scope 列表
      security:
        - ServiceToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserScopes"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{id}/roles:
    get:
      tags: [Roles]
      summary: 列出用户已分配角色
      security:
        - ServiceToken: []
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRoleList"
    post:
      tags: [Roles]
      summary: 为用户分配角色
      security:
        - ServiceToken: []
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role_id]
              properties:
                role_id:
                  type: string
      responses:
        "201":
          description: 分配成功
        "204":
          description: 重复分配（幂等）
        "422":
          description: role_type 不匹配（如给 User 分配 service_account 专属 Role）
  /users/{id}/roles/{role_id}:
    delete:
      tags: [Roles]
      summary: 移除用户角色
      security:
        - ServiceToken: []
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
        - name: role_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: 移除成功

  # ── Roles ──
  /roles:
    get:
      tags: [Roles]
      summary: 列出所有角色
      security:
        - ServiceToken: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
    post:
      tags: [Roles]
      summary: 创建自定义角色
      security:
        - ServiceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoleRequest"
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: 角色名已存在
...
  /service-tokens:
    get:
      tags: [System]
      summary: 列出所有 Service Token 元数据
      security:
        - ServiceToken: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceToken"
    post:
      tags: [System]
      summary: 创建 Service Token
      description: 为指定生态项目创建 Service Token，用于调用 /introspect
      security:
        - ServiceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project]
              properties:
                project:
                  type: string
                  enum: [pandaria, emerald, constell, tokencamp, heirloom]
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceToken"
        "409":
          description: 该项目已存在活跃 Service Token
...